Fixed major thread restarting issues + new page
For the CPT and VST, a new thread will be created and started every time they are ran. No longer need to restart the program to re-run a component! :)
Also added a new page tab for Thermal Conductivity Probe. Still just an empty page, will be adding more once the hardware is figured out

Temporarily commented out the actuator commands in the load cell reading function. Won't be using until actuator is finalized